Using Doubles for timelines from now on. Deltas still use floating-points for efficiency. * TimeTool.Quantize should now return the same value if called on the previous result of a similar quantize * CharacterMovement.OnValidate made overridable * Fixing seek flags being totally ignored in Reconcile

+++ Assets/Examples-Mirror/Scripts/B-TickerPhysics/PhysicsTickable.cs
@@ -11,7 +11,7 @@ namespace MultiplayerToolset.Examples.Mirror
 
         void PhysicsTick(float deltaTime, PhysicsPlayer.Input input);
 
-        PhysicsPlayer.Input GetInputAtTime(float time);
+        PhysicsPlayer.Input GetInputAtTime(double time);
     }
 
     public class PhysicsTickable : NetworkBehaviour, ITickable<PhysicsTickable.Input, PhysicsTickable.State>
